added authentication via JWT. started working on authorization

+++ src/controllers/accountController.js
@@ -1,8 +1,15 @@
 import { isEmpty } from "lodash";
 import Account from "../model/account";
 
+import { isAdmin } from "../helpers/authorizationHelper";
+
 class AccountController {
   async index(req, res) {
+    if (!isAdmin(req.user.email))
+      return res
+        .status(403)
+        .json({ message: "only admins are allowed to view all accounts" });
+
     const accounts = await Account.getAccounts();
     res.json({ accounts });
   }
@@ -15,6 +22,20 @@ class AccountController {
 
     const account = (await Account.getAccount({ id }))[0];
     if (account) {
+      console.log(
+        `${account.username}@${account.domain}`,
+        req.user.email,
+        `${account.username}@${account.domain}` !== req.user.email
+      );
+      if (
+        !isAdmin(req.user.email) &&
+        `${account.username}@${account.domain}` !== req.user.email
+      )
+        return res.status(403).json({
+          message:
+            "only admins and the user himself are allowed to view this account"
+        });
+
       res.json({ account });
     } else {
       res.status(404).json({ message: "account not found" });
@@ -30,6 +51,11 @@ class AccountController {
       enabled = true,
       sendonly = false
     } = req.body;
+
+    if (!isAdmin(req.user.email))
+      return res
+        .status(403)
+        .json({ message: "only admins are allowed to view all accounts" });
 
     if (!username || !domain || !password)
       return res.status(400).json({ message: "parameters are missing" });
@@ -55,6 +81,16 @@ class AccountController {
     const { username, domain, password, quota, enabled, sendonly } = req.body;
     const { id } = req.params;
 
+    const account = (await Account.getAccount({ id }))[0];
+    if (
+      !isAdmin(req.user.email) &&
+      `${account.username}@${account.domain}` !== req.user.email
+    )
+      return res.status(403).json({
+        message:
+          "only admins and the user himself are allowed to update this account"
+      });
+
     if (!req.body)
       return res.status(400).json({ message: "parameter are missing" });
     if (!id)
@@ -65,14 +101,20 @@ class AccountController {
       id
     );
 
-    const account = (await Account.getAccount({ id }))[0];
-    if (!account) return res.status(404).json({ message: "account not found" });
+    const updatedAccount = (await Account.getAccount({ id }))[0];
+    if (!updatedAccount)
+      return res.status(404).json({ message: "account not found" });
 
-    res.json({ account });
+    res.json({ updatedAccount });
   }
 
   async delete(req, res) {
     const { id } = req.params;
+
+    if (!isAdmin(req.user.email))
+      return res
+        .status(403)
+        .json({ message: "only admins are allowed to view all accounts" });
 
     const rowsDeleted = await Account.deleteAccount(id);
     if (rowsDeleted === 0)
